+++ react/lib/JBPasswordInput.tsx
@@ -14,11 +14,11 @@ const JBPasswordInput = React.forwardRef<JBPasswordInputWebComponent | undefined
 
   const element = useRef<JBPasswordInputWebComponent>(null);
   useImperativeHandle(ref, () => element.current ?? undefined, [element]);
-  const { onBeforeinput, onBlur, onChange, onEnter, onFocus, onInput, onKeydown, onKeyup, size, disabled, error, inputmode, label, message, name, placeholder, required, type, validationList, value, autocomplete, ...otherProps } = props;
+  const { onBeforeinput, onBlur, onChange, onEnter, onFocus, onInput, onKeydown, onKeyup, size, disabled, error, initialValue, inputmode, label, message, name, placeholder, required, type, validationList, value, autocomplete, ...otherProps } = props;
   // props that directly set in jsx dom and need no process or property set
-  const directProps: DirectProps = { label, message, name, placeholder, size, type, error, inputmode, autocomplete }
+  const directProps: DirectProps & { value: string, initialValue: string } = { label, message, name, placeholder, size, type, error, inputmode, autocomplete, value: value?.toString() ?? "", initialValue: initialValue?.toString() ?? "" }
   useJBInputEvents(element, { onBeforeinput, onBlur, onChange, onEnter, onFocus, onInput, onKeydown, onKeyup, ...otherProps });
-  useJBInputAttribute(element, { disabled, required, validationList, value, ...otherProps });
+  useJBInputAttribute(element, { disabled, required, validationList, ...otherProps });
   useEffect(() => {
     if (element.current) {
       element.current.minLength = props.minLength ?? null;
@@ -33,6 +33,7 @@ const JBPasswordInput = React.forwardRef<JBPasswordInputWebComponent | undefined
 
 export type Props = BaseProps<JBPasswordInputWebComponent> & {
   minLength?: number | null | undefined,
+  initialValue?: string | number | null,
 };
 JBPasswordInput.displayName = "JBPasswordInput";
 export { JBPasswordInput };