+++ CHANGELOG.md
@@ -4,4 +4,5 @@
 
 ### Changed
 
+- Added the React `initialValue` prop and forwarded `value` and `initialValue` directly as React 19 custom-element properties.
 - Breaking: renamed `--jb-password-input-active-eye-color` to `--jb-password-input-eye-color-active`.